requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
drf-spectacular==0.28.0
